+++ app/layout.tsx
@@ -3,16 +3,11 @@ import { Manrope } from 'next/font/google';
 import { Header } from '@/components/header';
 import { Footer } from '@/components/footer';
 import { FetchLayoutDocument, MenuItemFragment } from '@/generated/graphql';
-
 import { apiRequest } from '@/utils/api-request';
-import { CartProvider } from '@/context/cart-context';
-import { storage } from '@/core/storage.server';
-import { fetchCart } from '@/use-cases/fetch-cart';
-import { Cart } from '@/use-cases/contracts/cart';
-import { SidebarCart } from '@/components/sidebar-cart';
 
 import './globals.css';
-import { Suspense } from 'react';
+
+export const experimental_ppr = true;
 
 const manrope = Manrope({ subsets: ['latin'] });
 
@@ -40,21 +35,6 @@ type LayoutProps = { children: React.ReactNode };
 
 export default async function Layout({ children }: LayoutProps) {
     const { navigation } = await fetchLayout();
-    const getMyCart = async () => {
-        const cartId = await storage.getCartId();
-        const cart = cartId
-            ? await fetchCart(cartId)
-            : {
-                  items: [],
-                  total: {
-                      currency: 'eur',
-                      gross: 0,
-                      net: 0,
-                      taxAmount: 0,
-                  },
-              };
-        return cart as Cart;
-    };
 
     return (
         <html lang="en">
@@ -64,12 +44,9 @@ export default async function Layout({ children }: LayoutProps) {
                 <link rel="icon" href="/favicon.ico" />
             </head>
             <body className={`${manrope.className} bg-soft`}>
-                <CartProvider cartPromise={getMyCart()}>
-                    <Header navigation={navigation} />
-                    {children}
-                    <SidebarCart />
-                    <Footer navigation={navigation} />
-                </CartProvider>
+                <Header navigation={navigation} />
+                {children}
+                <Footer navigation={navigation} />
             </body>
         </html>
     );
