+++ components/cart-sidebar.tsx
@@ -1,20 +1,21 @@
 'use client';
 
 import Link from 'next/link';
-import { useCart } from '@/context/cart-context';
-import { Cart } from './cart';
-import { Price } from './price';
 import classNames from 'classnames';
+import { Cart } from '@/use-cases/contracts/cart';
 
-export const SidebarCart = () => {
-    const { cart, isCartOpen, setIsCartOpen } = useCart();
+import { CartItems } from './cart-items';
+import { Price } from './price';
 
+type SidebarCartProps = { cart: Cart | null; isOpen: boolean; onClose: () => void };
+
+export const CartSidebar = ({ cart, isOpen, onClose }: SidebarCartProps) => {
     return (
         <div
             className={classNames(
                 'bg-light transition-all border-l-muted border-l py-8 px-10 w-[500px] h-full z-10 overflow-y-auto',
                 'flex flex-col fixed top-0 right-0',
-                !isCartOpen && '!-right-[505px]',
+                !isOpen && '!-right-[505px]',
             )}
         >
             <div className="flex justify-between items-center mb-4">
@@ -22,13 +23,13 @@ export const SidebarCart = () => {
                 <button
                     aria-label="Close cart button"
                     className="bg-light h-8 w-8 font-bold border border-muted rounded hover:bg-muted/50"
-                    onClick={() => setIsCartOpen(false)}
+                    onClick={onClose}
                 >
                     x
                 </button>
             </div>
             <div className="grow h-full  overflow-y-scroll">
-                <Cart cart={cart} />
+                <CartItems cart={cart} />
             </div>
             <div>
                 {cart && cart.items.length > 0 && (
@@ -56,7 +57,7 @@ export const SidebarCart = () => {
                 <Link
                     href="/checkout"
                     className="bg-dark text-light text-lg rounded-xl px-8 py-4 w-full border border-10 block text-center"
-                    onClick={() => setIsCartOpen(false)}
+                    onClick={onClose}
                 >
                     Checkout
                 </Link>
